+++ etl/src/coacc_etl/ingest/custom/__init__.py
@@ -1,0 +1,1 @@
+"""Custom non-Socrata ingest adapters."""